weighing functions done

+++ database/migrations/2022_04_23_132059_create_results_table.php
@@ -16,22 +16,22 @@ class CreateResultsTable extends Migration
         Schema::create('results', function (Blueprint $table) {
             $table->id();
             $table->unsignedInteger('athlete_id')->index();
-            $table->integer('squat1')->nullable();
-            $table->integer('squat2')->nullable();
-            $table->integer('squat3')->nullable();
-            $table->integer('squat4')->nullable();
-            $table->integer('bench1')->nullable();
-            $table->integer('bench2')->nullable();
-            $table->integer('bench3')->nullable();
-            $table->integer('bench4')->nullable();
-            $table->integer('deadlift1')->nullable();
-            $table->integer('deadlift2')->nullable();
-            $table->integer('deadlift3')->nullable();
-            $table->integer('deadlift4')->nullable();
-            $table->integer('total')->nullable();
-            $table->integer('points')->nullable();
-            $table->integer('age_points')->nullable();
-            $table->integer('timbod')->nullable();         
+            $table->float('squat1')->nullable();
+            $table->float('squat2')->nullable();
+            $table->float('squat3')->nullable();
+            $table->float('squat4')->nullable();
+            $table->float('bench1')->nullable();
+            $table->float('bench2')->nullable();
+            $table->float('bench3')->nullable();
+            $table->float('bench4')->nullable();
+            $table->float('deadlift1')->nullable();
+            $table->float('deadlift2')->nullable();
+            $table->float('deadlift3')->nullable();
+            $table->float('deadlift4')->nullable();
+            $table->float('total')->nullable();
+            $table->float('points')->nullable();
+            $table->float('age_points')->nullable();
+            $table->float('timbod')->nullable();         
             $table->timestamps();
         });
     }
